move conversion stuff to submodule
