+++ generators/dotnet-oo-bindgen/src/conversion/mod.rs
@@ -1,0 +1,3 @@
+mod dotnet_type;
+
+pub(crate) use dotnet_type::*;
