+++ generators/dotnet-oo-bindgen/src/interface.rs
@@ -1,8 +1,8 @@
 use oo_bindgen::backend::*;
 use oo_bindgen::model::*;
 
+use crate::conversion::{base_functor_type, full_functor_type, DotnetType};
 use crate::doc::{docstring_print, xmldoc_print};
-use crate::dotnet_type::{base_functor_type, full_functor_type, DotnetType};
 use crate::formatting::{documentation, namespaced};
 use crate::helpers::call_dotnet_function;
 use crate::{print_imports, print_license};
